tgbot/requirements.txt
aiogram
python-dotenv
aiohttp
asyncio